Added some enumerations
We have a general idea of how to generate the level and rooms at this
point. Before we go too much further, the logic and graphics need to be
separated. The level generation logic should be done in the model while
the view should react to the model level definition and create the
graphical side of it. The next step is to move to a more MVP
architecture.