Added a simple graph implementation
Instead of defining the level as a 2-dimensional array, we will define
the level a graph where the vertices are rooms and the edges are the
tunnels between them. The 2-d array was a simple enough structure to
prove the concept but is not sufficient for the wide variety of items
that can exist in the level.