Adding types to help with path finding
At this point I am thinking to implement the A* path finding algorithm
to see what kind of dungeons can be generated. This needs some extra
types not supported in .Net 3.5 such a Tuple and PriorityQueue. These
will have basic implementations in order to get a working A* algorithm.
The focus is not on optimization.